+++ amqp-driver/src/main/scala/com/nummulus/amqp/driver/DefaultConsumer.scala
@@ -1,20 +1,27 @@
 package com.nummulus.amqp.driver
 
+import scala.concurrent.duration.Duration
+import scala.concurrent.duration.FiniteDuration
+
 import org.slf4j.LoggerFactory
 
 import com.nummulus.amqp.driver.IdGenerators._
 import com.nummulus.amqp.driver.akka.AkkaMessageConsumer
 import com.nummulus.amqp.driver.akka.AmqpQueueMessageWithProperties
 import com.nummulus.amqp.driver.api.consumer.AmqpConsumerRequest
+import com.nummulus.amqp.driver.api.consumer.AmqpConsumerRequestTimedOut
 import com.nummulus.amqp.driver.api.consumer.AmqpConsumerResponse
+import com.nummulus.amqp.driver.api.consumer.RequestTimedOut
 import com.nummulus.amqp.driver.configuration.QueueConfiguration
 
 import _root_.akka.actor.Actor
 import _root_.akka.actor.ActorRef
+import _root_.akka.actor.Scheduler
 
-class DefaultConsumer(
+private[driver] class DefaultConsumer(
     channel: Channel,
     configuration: QueueConfiguration,
+    timeOut: Duration,
     generateId: IdGenerator = IdGenerators.random) extends Actor {
   
   private val logger = LoggerFactory.getLogger(getClass)
@@ -29,7 +36,7 @@ class DefaultConsumer(
   channel.basicConsume(responseQueue, configuration.autoAcknowledge, generateId(), callback)
   
   private val pending = scala.collection.mutable.Map[String, Option[ActorRef]]()
-  
+
   def receive = {
     /**
      * Handles a request from another actor.
@@ -44,6 +51,16 @@ class DefaultConsumer(
       logger.debug("Sending message to queue: {}", body)
       logger.debug("Properties = {}", properties)
       channel.basicPublish("", configuration.queue, properties, body.getBytes)
+
+      if (sender.isDefined && timeOut.isFinite()) {
+        import context.dispatcher
+
+        logger.debug("Schedule time-out message to be sent in {}ms", timeOut.length)
+        scheduler.scheduleOnce(
+          timeOut.asInstanceOf[FiniteDuration],
+          self,
+          RequestTimedOut(properties.correlationId))
+      }
     }
     
     /**
@@ -65,5 +82,24 @@ class DefaultConsumer(
         logger.warn("Did not expect a response with correlationId {}", correlationId)
       }
     }
+
+    /**
+     * Handles a request time-out.
+     */
+    case RequestTimedOut(correlationId) => {
+      pending filterKeys (_ == correlationId) foreach {
+        case (_, Some(sender)) => sender ! AmqpConsumerRequestTimedOut
+        case (_, None) => // Response was received before the time-out
+      }
+
+      if (pending contains correlationId) {
+        pending -= correlationId
+      }
+    }
   }
+
+  /**
+   * Scheduler for time-out messages.
+   */
+  private[driver] def scheduler: Scheduler = context.system.scheduler
 }