+++ amqp-driver/src/main/scala/com/nummulus/amqp/driver/api/consumer/RequestTimedOut.scala
@@ -1,0 +1,3 @@
+package com.nummulus.amqp.driver.api.consumer
+
+private[driver] case class RequestTimedOut(correlationId: String)